loadMeasurementsForStreams function moved to ActiveSessionMeasurementsRepository
